// travelui/src/app/core/models/interface/flight.ts
export interface Flight {
    id: string;
    airline: string;
    airlineLogo: string;
    flightNumber: string;
    departureTime: string;
    arrivalTime: string;
    duration: string;
    price: number;
    stops: number;
    departureAirport: string;
    arrivalAirport: string;
    departureCity: string;
    arrivalCity: string;
}